Fixed various issues, reorganized docs folder, little cleanup of spheresvr.cpp.
- Fixed: WEIGHTREDUCTION not working correctly for items inside a container (Issue #289).
- Fixed: Retrieving a value from an unexistant LIST did not return -1; overwriting a LIST element generated an exception (Issue #290).
- Fixed: Profiler and worldsave message showing incorrect times (Issue #296).
- Fixed: Client listing showing the number of the online clients at the end of the list, instead of the top.
- Added: Doc file "Porting from 0.55 to 0.56.txt" (the content was posted years ago by XuN on the forum).

Changed GetPreciseSysTimeNano to GetPreciseSysTimeMicro, nanoseconds timer isn't so precise and it isn't even worth it.
Disabled DEBUGF_EXP log message when the delta exp is == 0.
Added class GlobalInitializer to initialize some global variables and do pre-server startup sanity checks and function calls.
Moved GetTimeMinDesc function from spheresvr.cpp to a static method of CServerTime.
Removed unused bFull parameter from CObjBase::CallPersonalTrigger.

+++ src/game/items/CItem.cpp
@@ -1863,7 +1863,7 @@ bool CItem::SetName( lpctstr pszName )
 int CItem::GetWeight(word amount) const
 {
 	int iWeight = m_weight * (amount ? amount : GetAmount());
-    int64 iReduction = GetPropNum(COMP_PROPS_ITEMCHAR, PROPITCH_WEIGHTREDUCTION, true);
+    int iReduction = GetPropNum(COMP_PROPS_ITEMCHAR, PROPITCH_WEIGHTREDUCTION, true);
 	if (iReduction)
 	{
 		iWeight -= (int)IMulDivLL( iWeight, iReduction, 100 );
